fix(desktop): collapse toolHint progress after final + e2e shared helpers
Three fixes:

1. removeTransientTurnMessagesSinceLastUser: mark retained toolHint
   progress with collapsed=true so they render collapsed after cleanup.

2. sendCleanup: stop calling cleanupListeners() when the typewriter
   finishes. Another final may arrive in the same turn. Listeners are
   now torn down only on abort/error/timeout/new-session.

3. onFinal: merge cleanup+dedup into single setMessages. Skip appending
   collapsed tool-call group when streaming already rendered toolHint
   progress to avoid duplicate entries.

e2e: extract shared Electron launch/bridge/helper into electron-setup.ts,
eliminating ~70 lines of beforeAll duplication per spec file.

Fixes 3 smoke tests in issue-172-assistant-turn-collapse.spec.ts.
